Add Artifactory support to deployment pipeline
This update introduces the ability to use Pypi/Maven packages from Artifactory for deployments, in addition to the existing DBFS file method. It includes new CLI options, code to handle authentication and upload to Artifactory, and updates to project configuration management. A test for the new functionality in the `deploy_v2` command is also provided.